Split to 2 mixins
